backend/requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Validation et modèles
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0

# Authentification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# XML
lxml==4.9.3

# XSLT
saxonche>=12.4.0

# Tâches en arrière-plan (à ajouter en Phase 3)
# celery==5.3.4
# redis==5.0.1

